split abstract auditing framework into separate module
